Completed Most Of The "UseItem" Code For Trimming Tool, 5/15/2023
Completed Most Of The "UseItem" Code For Trimming Tool. Took most of this code and stuff from the "Repair Tools" mod, with many alterations to make it work in this case. Also decided to name the current trimming tool "Graving Chisel", next I'll work on implemented the Graving Chisel, including sprite, and audio files, testing, etc, 5/15/2023.