Added "Graving Chisel" Tool & Did Some In-Game Testing, 5/16/2023
Added "Graving Chisel" Tool & Did Some In-Game Testing. Finally did some testing, and stuff seems to be working about as I expected, which is good to see. Now I mostly just need to add those audio-clips and implement them to play when the tool is used and such, do some more testing and such, and the mod should honestly be ready to be put public/released I think, cool stuff, 5/16/2023.