Commit Before Releasing v1.00, 5/16/2023
Commit Before Releasing v1.00. Forgot to do a commit for adding the audio-files and stuff, but that's alright. I did testing and everything seems to work, only annoying this is a quirk with equipping armor while another piece has just been unequipped in that spot. The bug made it so changing the variant while a piece was equipped would cause the armor values in that slot to get bugged out. So for now atleast I just said screw it and made it so the "new" piece does not auto-equip in those cases, just to prevent that bug from happening and causing worse issues than just a minor inconveniencing. Also will consider changing the Unity compression settings for the audio in future mods, the sizes are a bit excessive with the setting I just used for this and the last update for Repair Tools. But besides all that, I say this is ready for initial release, cool stuff, 5/16/2023.